added basic feedback form, tool tips,  logo inspection
Added a basic feedback form for the purpose of testing, I plan to update it to make it look nice in the soon future. Also added a tool tip to give a description of the logo that you are looking at for clubs with an animation to scale the logo up with glow around it when the user hovers over it.